requirements.txt
# Minimal dependencies for local development
sqlparse>=0.4.4
requests>=2.28
huggingface-hub>=0.15
transformers>=4.30
mysql-connector-python>=8.0
# Optional (for model loading/inference)
# torch>=2.0
